Converters done

+++ project-log/src/main/java/nz/co/farnault/projectlog/dto/MuscleGroupDTO.java
@@ -4,7 +4,7 @@ import java.util.Set;
 
 public class MuscleGroupDTO {
 
-	private int id;
+	private Long id;
 	private String name;
 	private Set<ExerciseDTO> exercises;
 
@@ -12,18 +12,18 @@ public class MuscleGroupDTO {
 		super();
 	}
 
-	public MuscleGroupDTO(int id, String name, Set<ExerciseDTO> exercises) {
+	public MuscleGroupDTO(Long id, String name, Set<ExerciseDTO> exercises) {
 		super();
 		this.id = id;
 		this.name = name;
 		this.exercises = exercises;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
